fix bug: return parsing not properly
There was an error in the return parsing (parsing, not translation), so I had to go back to the XML parser and fix it. Took me pretty much time. The bug was the incompatibility of 2 or more returns in the same statement block. Also, another bug with the return parsing was that the statement block must have ended with a return statement, which is not correct.